Remove unused variables; fix warnings about non-initialized vars; CONST-ified according to Tcl 8.4+ rules
bin/init.tcl: _ns_getscript forces import of
namespaced commands

tcl/init.tcl: sets auto_path to start with
our private library first

include/Makefile.global.in: allows for building
with Solaris 2.6 and later

+++ nscgi/nscgi.c
@@ -422,7 +422,7 @@ CgiInit(Cgi *cgiPtr, Map *mapPtr, Ns_Conn *conn)
 
     ulen = strlen(url);
     plen = strlen(mapPtr->url);
-    if ((strncmp(mapPtr->url, url, plen) == 0) &&
+    if ((strncmp(mapPtr->url, url, (size_t)plen) == 0) &&
     	(ulen == plen || url[plen] == '/')) {
 	
         if (mapPtr->path == NULL) {
@@ -588,7 +588,7 @@ CgiSpool(Cgi *cgiPtr, Ns_Conn *conn)
     fd = Ns_GetTemp();
     if (fd < 0) {
 	Ns_Log(Error, "nscgi: could not allocate temp file.");
-    } else if (write(fd, content, len) != len) {
+    } else if (write(fd, content, (size_t)len) != len) {
 	err = "write";
     } else if (lseek(fd, 0, SEEK_SET) != 0) {
 	err = "lseek";
